Phase2_Session10/requirements.txt
attrs==19.3.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
cycler==0.10.0
decorator==4.4.2
docutils==0.16
enum34==1.1.10
functools32==3.2.3.post2
future==0.18.2
futures==3.3.0
idna==2.9
importlib-metadata==1.6.0
ipython-genutils==0.2.0
jsonschema==3.2.0
jupyter-core==4.6.3
Kivy==1.11.1
kiwisolver==1.1.0
matplotlib==2.2.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbformat==4.4.0
numpy==1.16.5
olefile==0.46
opencv-contrib-python==4.2.0.32
opencv-python-headless==4.2.0.32
pathlib2==2.3.5
Pillow==6.2.0
plotly==3.10.0
pybullet==2.7.1
pycparser==2.19
pygame==1.9.6
Pygments==2.5.2
pyodbc==4.0.30
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
retrying==1.3.3
scandir==1.10.0
singledispatch==3.4.0.3
six==1.12.0
subprocess32==3.5.4
torch==1.4.0
tornado==5.1.1
traitlets==4.3.3
urllib3==1.25.8
zipp==1.2.0
